Remove subscription status from db table
It's redundant. We have the period start and end dates, trial period end
date and canceled_at date... The status of the subscription can be (and
now it is) obtained based on those attributes.

Previously to cancel a subscription we had to make two changes: update the
canceled_at date and explicily set the status to 'canceled'. Now it's only
matter of updating the canceled_at attribute and the status is
automatically set.